requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
future==0.18.2
idna==3.3
pytz==2021.3
requests==2.27.1
rets==1.0.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
xmltodict==0.12.0
